+++ Project.toml
@@ -1,0 +1,16 @@
+name = "APITemplate"
+uuid = "7bd4ac69-4294-43f9-8b92-99f6b4b0ee48"
+authors = ["Emirhan Kurtuluş <emirhan422@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
+JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+SimpleTokenizer = "65cef658-c627-4d23-8761-b7759d1bda69"
+Sockets = "6462fe0b-24de-5631-8697-dd941f90decc"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
