--- julia.packages
@@ -1,2 +1,0 @@
-HTTP@0.9
-JSON@0.21
