+++ tests/runtests.jl
@@ -1,0 +1,5 @@
+using Test
+using Pkg
+using SimpleTokenizer
+
+Pkg.test("SimpleTokenizer")
