--- vars.env
@@ -1,3 +1,0 @@
-export JULIA_REQ_FILE=julia.packages
-export API_PORT=5005
-export AUTH_TOKEN=sometoken282a916901834f
